Don't re-throw exceptions during init
Before, we were intentionally making init errors fatal as a signal to
drivers that something is wrong so they don't take the robot to the
field at competition if something needs to be fixed. We now use the
DriverInterface web page for this purpose, so we can be nicer about
reporting init errors, which is much better especially during initial
bringup of a robot.