add support for version controlled config files
optionally look for config files in the "config" subdir of the deploy directory, so they can be sourced and edited with the rest of the robot code.

the code will first look for a config file named "selectedConfig.txt" which is meant to be a symlink to one of potentially several config files in the repository.

if that file does not exist, the code will back off to look for "robotConfig.txt" under the deploy directory.

if that file does not exist, the code will back off to the legacy config file location under "/home/lvuser".